+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-4.9.3 / 5.2.3 (in progress)
+4.9.4 / 5.2.4 (in progress)
+================
+* [#1320](https://github.com/oshi/oshi/pull/1320): Avoid NPE for failed Windows registry counter size - [@dbwiddis](https://github.com/dbwiddis).
+* Your contribution here
+
+4.9.3 / 5.2.3 (2020-08-09)
 ================
 * [#1310](https://github.com/oshi/oshi/pull/1310): Use fragment size in Linux FileStore size calculation - [@dbwiddis](https://github.com/dbwiddis).
 * [#1316](https://github.com/oshi/oshi/pull/1316): Fix ARM architecture parsing - [@dbwiddis](https://github.com/dbwiddis).